finish ex 14

+++ exercises/Chapter2/src/ex14/powerOfTwo.cc
@@ -5,7 +5,7 @@ using namespace std;
 int main() 
 {
     
-    int x;
+    size_t x;
     cin >> x;
     
     bool powerOfTwo =  x != 0 && (x & (x - 1)) == 0;
@@ -13,4 +13,5 @@ int main()
     cout << (powerOfTwo ? "The value is an exact power of two" : 
                           "The value is not an exact power of two" ) 
                           << '\n';   
+                          
 }
